+++ CHANGES.rst
@@ -4,7 +4,9 @@ OpenREM version history
 
 1.0.0b1 (2020-xx-xx)
 --------------------
+* `#818`_  Interface: refactored event number filtering
 * `#816`_  Interface: fixed password change error
+* `#815`_  Interface: fixed patient name filtering
 * `#814`_  Deployment: automated deployment to dev.openrem.org and testing.openrem.org reintroduced
 * `#808`_  Imports: caters for illegal use of mGy units in RDSR for dose at RP values
 * `#807`_  Exports: fixed errors in PHE fluoro export when values are None